requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
PyMySQL>=1.0
psycopg2-binary>=2.9
python-dotenv>=1.0